# gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Shopping
  description: 在线商城服务
  version: "1.0"
host: localhost:8080
consumes:
- application/json
produces:
- application/json
paths:
...
  /shop:
    get:
      tags:
      - shop
      summary: 分组列表
      description: |-
        商品列表

        **Required security scopes for jwt**:
          * `role:user`
          * `role:admin`
      operationId: shop#GoodsList
      parameters:
      - name: cursor
        in: query
        description: cursor of page
        required: true
        type: integer
      - name: limit
        in: query
        description: limit of items
        required: true
        type: integer
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ShopGoodsListResponseBody'
            required:
            - items
            - nextCursor
            - total
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/ShopGoodsListBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ShopGoodsListInternalServerErrorResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
  /shop/{goodsId}:
    get:
      tags:
      - shop
      summary: 商品详情
      description: |-
        商品详情

        **Required security scopes for jwt**:
          * `role:user`
          * `role:admin`
      operationId: shop#GetGoods
      produces:
      - application/json
      parameters:
      - name: goodsId
        in: path
        required: true
        type: string
        minLength: 1
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: true
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ShopGetGoodsResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/ShopGetGoodsBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ShopGetGoodsInternalServerErrorResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
    put:
      tags:
      - shop
      summary: 更新商品
      description: |-
        更新商品

        **Required security scopes for jwt**:
          * `role:admin`
      operationId: shop#UpdateGoods
      produces:
      - application/json
      parameters:
      - name: goodsId
        in: path
        required: true
        type: string
        minLength: 1
      - name: Authorization
        in: header
        description: JWT used for authentication
        required: true
        type: string
      - name: UpdateGoodsRequestBody
        in: body
        required: true
        schema:
          $ref: '#/definitions/ShopUpdateGoodsRequestBody'
          required:
          - name
          - description
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/ShopUpdateGoodsResponseBody'
        "400":
          description: Bad Request response.
          schema:
            $ref: '#/definitions/ShopUpdateGoodsBadRequestResponseBody'
        "500":
          description: Internal Server Error response.
          schema:
            $ref: '#/definitions/ShopUpdateGoodsInternalServerErrorResponseBody'
      schemes:
      - http
      security:
      - jwt_header_Authorization: []
...
definitions:
...
  GoodsBrandResponseBody:
    title: 'Mediatype identifier: goodsbrand; view=default'
    type: object
    properties:
      description:
        type: string
        description: 品牌描述
        example: 耐克描述
      id:
        type: integer
        description: ID
        example: 7487309560206597432
        format: int64
      name:
        type: string
        description: 品牌名称
        example: 耐克
    description: 商品品牌
    example:
      description: 耐克描述
      id: 2275386429741612544
      name: 耐克
    required:
    - id
    - name
    - description
...
  GoodsSkuResponseBody:
    title: 'Mediatype identifier: goodssku; view=default'
    type: object
    properties:
      goods_sku_value:
        type: array
        items:
          $ref: '#/definitions/GoodsSkuValueResponseBody'
        description: 商品SKU规格值
        example:
        - id: 322396054482440455
          name: 红色
        - id: 322396054482440455
          name: 红色
        - id: 322396054482440455
          name: 红色
        - id: 322396054482440455
          name: 红色
      id:
        type: integer
        description: ID
        example: 1325989832679285006
        format: int64
      name:
        type: string
        description: 颜色
        example: 颜色
    description: 商品SKU
    example:
      goods_sku_value:
      - id: 322396054482440455
        name: 红色
      - id: 322396054482440455
        name: 红色
      - id: 322396054482440455
        name: 红色
      - id: 322396054482440455
        name: 红色
      id: 1883767716292753139
      name: 颜色
    required:
    - id
    - name
  GoodsSkuValueResponseBody:
    title: 'Mediatype identifier: goodsskuvalue; view=default'
    type: object
    properties:
      id:
        type: integer
        description: ID
        example: 3185204166962498619
        format: int64
      name:
        type: string
        description: 红色
        example: 红色
    description: 商品SKU规格值
    example:
      id: 6445452259142062634
      name: 红色
    required:
    - id
    - name
...
  GoodsTypeResponseBody:
    title: 'Mediatype identifier: goodstype; view=default'
    type: object
    properties:
      description:
        type: string
        description: 类型描述
        example: 男人的衣柜
      id:
        type: integer
        description: ID
        example: 2387300402565235828
        format: int64
      name:
        type: string
        description: 类型名称
        example: 男装
    description: 商品分类
    example:
      description: 男人的衣柜
      id: 8583831439453180556
      name: 男装
    required:
    - id
    - name
    - description
...
  ShopGetGoodsBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: GetGoods_bad_request_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ShopGetGoodsInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: GetGoods_internal_server_error_Response_Body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ShopGetGoodsResponseBody:
    title: 'Mediatype identifier: goodsspu; view=default'
    type: object
    properties:
      description:
        type: string
        description: 商品描述
        example: 水果机是世界上最好的手机
      goods_brand:
        $ref: '#/definitions/GoodsBrandResponseBody'
      goods_sku:
        type: array
        items:
          $ref: '#/definitions/GoodsSkuResponseBody'
        description: 商品SKU
        example:
        - goods_sku_value:
          - id: 322396054482440455
            name: 红色
          - id: 322396054482440455
            name: 红色
          id: 3278873586745833823
          name: 颜色
        - goods_sku_value:
          - id: 322396054482440455
            name: 红色
          - id: 322396054482440455
            name: 红色
          id: 3278873586745833823
          name: 颜色
        - goods_sku_value:
          - id: 322396054482440455
            name: 红色
          - id: 322396054482440455
            name: 红色
          id: 3278873586745833823
          name: 颜色
        - goods_sku_value:
          - id: 322396054482440455
            name: 红色
          - id: 322396054482440455
            name: 红色
          id: 3278873586745833823
          name: 颜色
      goods_type:
        $ref: '#/definitions/GoodsTypeResponseBody'
      id:
        type: integer
        description: ID
        example: 1721886835493197713
        format: int64
      name:
        type: string
        description: 商品名称
        example: 水果机
    description: GetGoodsResponseBody result type (default view)
    example:
      description: 水果机是世界上最好的手机
      goods_brand:
        description: 耐克描述
        id: 1579160059629799796
        name: 耐克
      goods_sku:
      - goods_sku_value:
        - id: 322396054482440455
          name: 红色
        - id: 322396054482440455
          name: 红色
        id: 3278873586745833823
        name: 颜色
      - goods_sku_value:
        - id: 322396054482440455
          name: 红色
        - id: 322396054482440455
          name: 红色
        id: 3278873586745833823
        name: 颜色
      goods_type:
        description: 男人的衣柜
        id: 7133122405476397543
        name: 男装
      id: 4838685326929744627
      name: 水果机
    required:
    - id
    - name
    - description
    - goods_type
    - goods_brand
  ShopGoodsListBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: true
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: true
    description: GoodsList_bad_request_Response_Body result type (default view)
    example:
      fault: false
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ShopGoodsListInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: GoodsList_internal_server_error_Response_Body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: false
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
...
  ShopUpdateGoodsBadRequestResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: false
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: UpdateGoods_bad_request_Response_Body result type (default view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: false
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ShopUpdateGoodsInternalServerErrorResponseBody:
    title: 'Mediatype identifier: application/vnd.goa.error; view=default'
    type: object
    properties:
      fault:
        type: boolean
        description: Is the error a server-side fault?
        example: false
      id:
        type: string
        description: ID is a unique identifier for this particular occurrence of the
          problem.
        example: 123abc
      message:
        type: string
        description: Message is a human-readable explanation specific to this occurrence
          of the problem.
        example: parameter 'p' must be an integer
      name:
        type: string
        description: Name is the name of this class of errors.
        example: bad_request
      temporary:
        type: boolean
        description: Is the error temporary?
        example: true
      timeout:
        type: boolean
        description: Is the error a timeout?
        example: false
    description: UpdateGoods_internal_server_error_Response_Body result type (default
      view)
    example:
      fault: true
      id: 123abc
      message: parameter 'p' must be an integer
      name: bad_request
      temporary: true
      timeout: true
    required:
    - name
    - id
    - message
    - temporary
    - timeout
    - fault
  ShopUpdateGoodsRequestBody:
    title: ShopUpdateGoodsRequestBody
    type: object
    properties:
      description:
        type: string
        description: 商品描述
        example: 水果x刘海屏就很丑
        maxLength: 128
      name:
        type: string
        description: 商品名称
        example: 水果x
        maxLength: 20
    example:
      description: 水果x刘海屏就很丑
      name: 水果x
    required:
    - name
    - description
  ShopUpdateGoodsResponseBody:
    title: 'Mediatype identifier: goodsspu; view=default'
    type: object
    properties:
      description:
        type: string
        description: 商品描述
        example: 水果机是世界上最好的手机
      goods_brand:
        $ref: '#/definitions/GoodsBrandResponseBody'
      goods_sku:
        type: array
        items:
          $ref: '#/definitions/GoodsSkuResponseBody'
        description: 商品SKU
        example:
        - goods_sku_value:
          - id: 322396054482440455
            name: 红色
          - id: 322396054482440455
            name: 红色
          id: 3278873586745833823
          name: 颜色
        - goods_sku_value:
          - id: 322396054482440455
            name: 红色
          - id: 322396054482440455
            name: 红色
          id: 3278873586745833823
          name: 颜色
      goods_type:
        $ref: '#/definitions/GoodsTypeResponseBody'
      id:
        type: integer
        description: ID
        example: 113221754232817214
        format: int64
      name:
        type: string
        description: 商品名称
        example: 水果机
    description: UpdateGoodsResponseBody result type (default view)
    example:
      description: 水果机是世界上最好的手机
      goods_brand:
        description: 耐克描述
        id: 1579160059629799796
        name: 耐克
      goods_sku:
      - goods_sku_value:
        - id: 322396054482440455
          name: 红色
        - id: 322396054482440455
          name: 红色
        id: 3278873586745833823
        name: 颜色
      - goods_sku_value:
        - id: 322396054482440455
          name: 红色
        - id: 322396054482440455
          name: 红色
        id: 3278873586745833823
        name: 颜色
      - goods_sku_value:
        - id: 322396054482440455
          name: 红色
        - id: 322396054482440455
          name: 红色
        id: 3278873586745833823
        name: 颜色
      goods_type:
        description: 男人的衣柜
        id: 7133122405476397543
        name: 男装
      id: 3465519874152322210
      name: 水果机
    required:
    - id
    - name
    - description
    - goods_type
    - goods_brand
...
securityDefinitions:
  jwt_header_Authorization:
    type: apiKey
    description: |-
      JWT 认证

      **Security Scopes**:
        * `role:user`: 普通用户
        * `role:admin`: 管理员
    name: Authorization
    in: header